+++ packages/grid-export/src/common/CellFormatters.ts
@@ -1,0 +1,5 @@
+export type CustomCellFormatters<TItem, TColumnKey extends string> = Partial<
+  Record<TColumnKey, CustomCellFormatter<TItem>>
+>;
+
+export type CustomCellFormatter<TItem> = (item: TItem) => string | number;