+++ packages/grid-export/src/features/grid-copy-to-clipboard/transformers/__tests__/CellTransformer.test.ts
@@ -4,61 +4,46 @@ describe("CellTransformer", () => {
   describe("transformItemToCell", () => {
     describe("when there is no itemLabelFormatter", () => {
       describe("and value is string", () => {
-        it("returns value and type string", () => {
+        it("returns value", () => {
           expect(
             transformItemToCell({ x: "hej" }, (item) => item.x, undefined)
-          ).toStrictEqual({
-            value: "hej",
-            type: "string",
-          });
+          ).toStrictEqual("<td>hej</td>");
         });
       });
       describe("and value is number", () => {
-        it("returns value and type number", () => {
+        it("returns value", () => {
           expect(
             transformItemToCell({ x: 123 }, (item) => item.x, undefined)
-          ).toStrictEqual({
-            value: 123,
-            type: "number",
-          });
+          ).toStrictEqual("<td>123</td>");
         });
       });
       describe("and value is boolean", () => {
         describe("and value is true", () => {
-          it("returns Y and type string", () => {
+          it("returns Y ", () => {
             expect(
               transformItemToCell({ x: true }, (item) => item.x, undefined)
-            ).toStrictEqual({
-              value: "Y",
-              type: "string",
-            });
+            ).toStrictEqual("<td>Y</td>");
           });
         });
         describe("and value is false", () => {
-          it("returns empty and type string", () => {
+          it("returns empty", () => {
             expect(
               transformItemToCell({ x: false }, (item) => item.x, undefined)
-            ).toStrictEqual({
-              value: "",
-              type: "string",
-            });
+            ).toStrictEqual("<td></td>");
           });
         });
       });
       describe("and value is Date", () => {
-        it("returns formatted date and type string", () => {
-          let date = new Date(2020, 4, 9, 12, 0, 0);
+        it("returns formatted date", () => {
+          const date = new Date(2020, 4, 9, 12, 0, 0);
           expect(
             transformItemToCell({ x: date }, (item) => item.x, undefined)
-          ).toStrictEqual({
-            value: "2020-05-09 12:00",
-            type: "string",
-          });
+          ).toStrictEqual("<td>2020-05-09 12:00</td>");
         });
       });
     });
     describe("when there is an itemLabelFormatter", () => {
-      it("returns label and type string", () => {
+      it("returns label", () => {
         let date = new Date(2020, 4, 9, 12, 0, 0);
         expect(
           transformItemToCell(
@@ -66,15 +51,12 @@ describe("CellTransformer", () => {
             (item) => item.x,
             () => "hello"
           )
-        ).toStrictEqual({
-          value: "hello",
-          type: "string",
-        });
+        ).toStrictEqual("<td>hello</td>");
       });
     });
     describe("with custom format", () => {
       describe("custom format is string", () => {
-        it("returns custom format and type string", () => {
+        it("returns custom format", () => {
           let date = new Date(2020, 4, 9, 12, 0, 0);
           expect(
             transformItemToCell(
@@ -83,14 +65,11 @@ describe("CellTransformer", () => {
               undefined,
               () => "custom"
             )
-          ).toStrictEqual({
-            value: "custom",
-            type: "string",
-          });
+          ).toStrictEqual("<td>custom</td>");
         });
       });
       describe("custom format is number", () => {
-        it("returns custom format and type number", () => {
+        it("returns custom format", () => {
           let date = new Date(2020, 4, 9, 12, 0, 0);
           expect(
             transformItemToCell(
@@ -99,10 +78,7 @@ describe("CellTransformer", () => {
               undefined,
               () => 3
             )
-          ).toStrictEqual({
-            value: 3,
-            type: "number",
-          });
+          ).toStrictEqual("<td>3</td>");
         });
       });
     });